+++ client/src/App.tsx
@@ -6,6 +6,7 @@ import AddBook from "./pages/AddBook";
 import Browse from "./pages/Browse";
 import Stats from "./pages/Stats";
 import ResetPassword from "./pages/ResetPassword";
+import Account from "./pages/Account";
 
 const App: React.FC = () => {
   return (
@@ -15,6 +16,7 @@ const App: React.FC = () => {
         <Route path="/add" Component={AddBook} />
         <Route path="/browse" Component={Browse} />
         <Route path="/stats" Component={Stats} />
+        <Route path="/account" Component={Account} />
         <Route path="/resetpassword" Component={ResetPassword} />
         <Route path="*" Component={NotFound} />
       </Routes>